Move Dashboard into DashboardLayout
